[REFACTOR] Add Protocol Folder & Script
